+++ integrations/github/github/clients/auth/github_app/installation_authenticator.py
@@ -15,7 +15,7 @@ class GitHubAppInstallationAuthenticator(AbstractGitHubAuthenticator):
         self,
         app_auth: GitHubAppAuthenticator,
         organization: str,
-        installation_id: Optional[str] = None,
+        installation_id: str,
     ):
         self.app_auth = app_auth
         self.organization = organization
@@ -25,8 +25,7 @@ class GitHubAppInstallationAuthenticator(AbstractGitHubAuthenticator):
 
     @property
     def rate_limit_scope(self) -> str:
-        # TODO: replace to installation id based rate limit scope after create_github_client is deprecated
-        return f"installation:{self.organization}"
+        return f"installation:{self.installation_id}"
 
     async def get_headers(self) -> GitHubHeaders:
         token_response = await self.get_token()